fix OpenGL artefacts at scene export
When exporting the scene for the second time, weird artefacts that looked
like shadows occured. The fix enforces to re-allocate the resources of
the offscreen QOpenGLContext. Measurements showed that reallocation is
much faster than the rate-limiting step of the Export function: dump the
file to disk. Performance impact is therefore limited. Feel free to
investigate and propose a more elegant fix...