lDDT: limit max number of possible chain mappings
Not that it is helpful to enumerate more chain mappings than atoms
in the universe, but the limit is mainly in place to avoid a
memory explosion. We're using a itertools.product there. From the
docu: Before product() runs, it completely consumes the input
iterables, keeping pools of values in memory to generate the products.
Accordingly, it is only useful with finite inputs.