Accept 0.0 QS-score mappings in QS-score based optimization
The reason for that is that one might start with an initial seed
somewhere in Nirvana that leads to no possible increase in QS-score.
Stoping there might hinder a non-zero QS-score when continuing with
other seeds. And even if the QS-score is 0.0 in the end, the mapping
should make sense from an lDDT perspective more or less...