chain mapping: Extend processing of input structures
All structures, be it target or model structures undergo additional
processing so they only contain residues with complete backbone.
For peptides thats: N, CA, C, CB (no CB for GLY), for nucleotides
thats O5', C5', C4', C3' and O3'. Reason for that is to give some
guarantees to downstream algorithms.

One downstream algorithm that could be adapted because of that is
the transformation in the ReprResult class. If number of residues
is smaller than 3, all those backbone atoms are used to derive
a transformation in order to not crash. Behaviour for ReprResult
objects with >= 3 residues does not change. Only CA/C3' atoms
are used.