Correctly read/write CONECT in PDB files - Thanks Simon Sun
Simon found issues when reading and writing CONECT statements
in PDB io and provided code to fix it. yay

Reading: OpenStructure did not read CONECT statements at all. By default,
it still doesnt. Many PDB files out there don't provide these
statements and we really don't want to rely on them. We rather want
to assign connectivity based on the chemical component dictionary from
the PDB. However, a valid use case are novel compounds that are not
in the component dictionary. Reading connect statements can now be
enabled in the pdb reader through the IOProfile. This may give issues
in processing after reading. OpenStructure implements processors that
are responsible for connectivity. Now that we build some of the
connectivity already at the reading stage, this might cause trouble.
To remedy most of the nightmares that can come out of that, the
processors can now optionally skip connectivities between Hetatoms.

Writing: That was a plain bug when writing CONECT statements for
bond orders > 1.