be more permissive when reading atom serial number
OST writes invalid atom serial numbers if they exceed the PDB format limits.
That's a feature not a bug. The new behaviour is that invalid serial numbers
don't get parsed. Any processing function working with these serial numbers
need to manually check for that. The only current effect is when reading
CONECT statements. Atoms with invalid serial numbers are not connected.