fix: remove H with selection rather than Molck
This makes the ligand scoring script more independent of an up-to-date
compound library, which is a strong requirement of Molck.
This was a problem for the April 2023 Modeling Championship where the
script was run with an out of date library and the element of an atom of
the ligand was silently changed, preventing the symmetry correction to
find isomorphisms and resulting in the absence of scores.