Enable global chain mappings in ost.mol.alg.ChainMapper.GetRepr
GetRepr identifies the best possible representation of a reference
substructre in a model including chain mapping etc. It also enriches
it with attributes/scores that are useful to analyse binding sites etc.
Enforcing a global mapping fully defines the desired representation in
the model but extracts it and enriches it with all the nice
attributes of the GetRepr output.