Scoring: Use consistent definition of n_max_naive parameter
There are two use cases for this parameter: 1) decide whether a chain
mapping problem is actually solvable by naive enumeration and crash out if not
2) As a decision whether one should use naive enumeration or a heuristic
approach for chain mapping.

For the latter, n_max_naive meant the number of chains in the underlying
structure. Now its number of possible mappings for both cases.