Revert "avoid atan2(0.0, 0.0) in Dihedral::SetAngleICS"
This reverts commit 9153648f1b773a1c1186d0a0ac29e78b659cd1f8.

The intention of this commit was to fix potential issues when operating on a
dihedral connecting 4 atoms where at least 3 atoms are exactly on a line.
Dihedrals are ill defined in this case but "expected behaviour"
(whatever that means) is tested as a unit test.
A fail was observed in two cases. First case that triggered commit
9153648f1b773a1c1186d0a0ac29e78b659cd1f8, second case that showed that the
problem is not solved. Therefore reverting to original state in order to find
different solution.