Disable unit test for internal coordinates edge case
This unit test has been observed to fail (EXTREMELY RARELY)
It tests the edge case of 3 atoms that are all on a line.
From an external coordinate system perspective, the dihedral
is not defined in this case.
From an internal coordinate perspective this should not matter.
The last line is the failing line. The expected behaviour is
that a rotation of x around t1 (The "linear" dihedral) should
propagate and affect t2 accordingly. Observed failure:
difference{4} between s.t2.GetAngle(){-2.3561945} and Real(3.14159265358979323846/4){0.785398185} exceeds 0.0001%

In the meantime we propose to use internal coordinates with caution and
open a task in JIRA for someone brave enough to deep dive into the code.