mmcif writing: make mmCIFify optional
There are two modes of mmCIF writing.
- Entities that are already mmCIF conform: Water/ligands are in
  separate chains etc.
- Entities are a mess: Chains must be split and we need to
  assign educated guesses of chain types.
The writer currently implements the second scenario. This commit introduces
a flag to choose between the two scenarios. In case of the first scenario
there is not much guessing but rather sanity checks whether the residues
of the respective chains are compatible with the chain types that are set.
The code itself is not working yet.