mmcif writer: resnum based entity matching, even if atomseq differs
By default, the mmcif_conform flag is on. That means, that we assume
residue numbers to be correctly set according to the underlying SEQRES.
This allows for residue number based alignment between chains.
BUT: For the scenario of no user defined mmcif entities (which is not possible
as of now anyways), there is no other way than deriving entities and
their SEQRES from the structure itself. Let's assume an Entity with
SEQRES HELLYEAH and two chains with ATOMSEQ HELLYEA and HELLYEAH.

The new implementation iterates over the first chain and defines an
entity with SEQRES HELLYEA. It then process the second chain and sees
that we have a perfect match but an added residue. If the fraction of
added residues with respect to matches is below a certain threshold,
the entity SEQRES gets updated and the second chain gets also assigned
to that entity.