mmcif writer: resolve linking errors for certain usages of ChemClass
Assignment of ChemClass values happens inside the ChemClass definition
in the header file. Let's say you're using something like:
ChemClass::WATER. If this is resolved at compile time, you won't have
any problem. If not, the linker searches for it in the mol shared object
file. BUT ITS NOT THERE. The hope is that defining the types as an
enum inside the ChemClass definition helps.