biounits: naming conflicts of AU chains with identity transform applied
Chain naming in biounit construction reserves 1.<au_cname> for the
respective AU chain with identity transform applied. As of January
2024, there are 3 PDB entries (8qn6, 8x1h, 2c0x) where the same AU
chain with identity transform occurs several times in the same
biounit. This is likely an error in the respective mmCIF files as
the resulting chains sit on top of each other. OST just names the
FIRST occurence as 1.<au_cname>.

A warning for that behaviour is also added to the documentation