mmcif writer: Heuristic for different _chem_comp.type for same compound name
The MMCifWriter stores globally the types of all residues in a structure.
We do a consistency check that the same residue name always has the same
type. However, there were special cases that we observed in the PDB. One
example is entry 2YHX which has plenty of UNK. UNK is 'L-PEPTIDE LINKING'
according to components dictionary. However, OST reads that info from the
components dictionary AND checks whether there are additional unknown atoms.
If yes, ChemClass is set to UNKNOWN which results in _chem_comp.type OTHER.
In the case of 2YHX we now have a mix of "L-PEPTIDE-LINKING" and "OTHER".

The added heuristic allows such a mix of two type, if one of the types is
"OTHER" and then stores the one that is NOT "OTHER". In this case
"L-PEPTIDE-LINKING"