feat: limit number of symmetries
Some ligands like YUT (in CAMEO 2023-12-30_00000023/8G9C) have a huge
number of isomorphic mappings and cause CAMEO to run out of time in the
scoring (and fail horribly). This allows to limit how much time is spent
listing all the isomorphisms. 100000 (1e5) seems reasonable and ensures
the scoring completes in minutes. Computing RMSD with SCRMSD is a bit
faster and 1000000 (1e6) should be OK.

New unassignment reasons are introduced as well.