mmcif writer: Derive info on chemical components from compound library
previously this was done with residue ChemClasses. This is set by the
compound library anyways during processing when loading a structure.
The problem: when we suddenly need info on chemical components for stuff
that is in the SEQRES but nowhere resolved in the structure. I really
tried hard to avoid the requirement for the compound library in mmcif
writing but there seems no way around.