fix: extract ligands from IsLigand flag
It turns out for some reason we were reading them from chains, based on
the ChainType of the chain. This prevented extracting anything from
legacy PDB files and didn't match the documented behavior in the action.

Now extraction is based on the IsLigand flag only, and documentation is
consistent throughout. Behavior on PDB files is tested.